general: unclutter App – inline IndexWithProviders
